realestate/requirements.txt
Django==1.9.5
djangorestframework==3.3.3
model-mommy==1.2.6
psycopg2==2.6.1
py-moneyed==0.6.0
dj-database-url==0.3.0
python-decouple==3.0
gunicorn==19.4.5
us==0.9.1
django-bootstrap-form==3.2
django-compressor==2.0
django-appconf==1.0.2
html5lib==0.9999999
slimit==0.8.1
lxml==3.6.0
django-redis-cache==1.6.5
hiredis==0.2.0
django-htmlmin==0.9.1
redis==2.10.5
requests==2.10.0
Pillow==3.2.0
django-storages==1.4.1
boto==2.40.0